refactor(common): one string dictionary for both state planes
