+++ packages/lix/src/common/mod.rs
@@ -5,6 +5,7 @@ pub(crate) mod identity;
 pub(crate) mod json_pointer;
 pub(crate) mod lix_path;
 pub(crate) mod metadata;
+pub(crate) mod string_dictionary;
 pub(crate) mod timestamp;
 pub(crate) mod types;
 pub(crate) mod wire;
@@ -21,6 +22,9 @@ pub use lix_path::{LixPath, validate_lix_path_segment};
 pub(crate) use lix_path::{compose_directory_path, compose_file_path};
 pub(crate) use metadata::{
     parse_row_metadata, parse_row_metadata_value, serialize_row_metadata, validate_row_metadata,
+};
+pub(crate) use string_dictionary::{
+    FastHashBuilder, StringDictionary, StringDictionaryBuilder, fast_hash_builder,
 };
 pub(crate) use timestamp::LixTimestamp;
 pub use types::{Blob, Json, LixNotice, NullableKeyFilter, SharedStr, SqlQueryResult, Value};
