+++ packages/lix/src/tracked_state/row_materialization.rs
@@ -1,50 +1,19 @@
 use std::collections::{BTreeMap, HashMap};
 #[cfg(test)]
 use std::mem::size_of;
+#[cfg(test)]
 use std::ops::Range;
-
-use ahash::RandomState;
-use bytes::Bytes;
 
 use crate::LixError;
 use crate::changelog::{
     ChangeId, ChangeRecordProjection, CommitId, MaterializedChangePayload,
     materialize_known_change_payloads,
 };
-use crate::common::{LixTimestamp, SharedStr};
+use crate::common::{LixTimestamp, SharedStr, StringDictionary, StringDictionaryBuilder};
 use crate::entity_pk::EntityPk;
 use crate::storage_adapter::StorageAdapterRead;
 use crate::tracked_state::MaterializedTrackedStateRow;
 use crate::tracked_state::types::{TrackedStateIndexValue, TrackedStateKey, TrackedStateKeyRef};
-
-#[derive(Debug, Default)]
-struct TrackedStateStringDictionary {
-    bytes: Bytes,
-    ranges: Vec<Range<u32>>,
-    #[cfg(test)]
-    arena_allocation_count: usize,
-    #[cfg(test)]
-    arena_large_allocation_count: usize,
-}
-
-impl TrackedStateStringDictionary {
-    fn get(&self, ordinal: u32) -> &str {
-        let range = self
-            .ranges
-            .get(ordinal as usize)
-            .expect("tracked-state string ordinal belongs to this batch");
-        let range = range.start as usize..range.end as usize;
-        // SAFETY: the builder appends complete `str` values and records their
-        // exact boundaries in one immutable allocation.
-        unsafe { std::str::from_utf8_unchecked(&self.bytes[range]) }
-    }
-
-    fn shared(&self, ordinal: u32) -> SharedStr {
-        let value = self.get(ordinal);
-        SharedStr::from_utf8_slice(self.bytes.clone(), value)
-            .expect("tracked-state dictionary value points into its byte arena")
-    }
-}
 
 #[derive(Debug)]
 struct MaterializedTrackedStateDescriptor {
@@ -68,7 +37,7 @@ struct MaterializedTrackedStateDescriptor {
 /// The legacy owned row is constructed only by an explicit terminal adapter.
 #[derive(Debug, Default)]
 pub(crate) struct MaterializedTrackedStateBatch {
-    strings: TrackedStateStringDictionary,
+    strings: StringDictionary,
     rows: Vec<MaterializedTrackedStateDescriptor>,
 }
 
@@ -136,15 +105,15 @@ impl MaterializedTrackedStateBatch {
 
     #[cfg(test)]
     pub(crate) fn dictionary_entry_count(&self) -> usize {
-        self.strings.ranges.len()
+        self.strings.len()
     }
 
     #[cfg(test)]
     pub(crate) fn large_buffer_count(&self, threshold: usize) -> usize {
         [
             self.rows.capacity() * size_of::<MaterializedTrackedStateDescriptor>(),
-            self.strings.bytes.len(),
-            self.strings.ranges.len() * size_of::<Range<u32>>(),
+            self.strings.byte_len(),
+            self.strings.len() * size_of::<Range<u32>>(),
         ]
         .into_iter()
         .filter(|bytes| *bytes >= threshold)
@@ -153,12 +122,12 @@ impl MaterializedTrackedStateBatch {
 
     #[cfg(test)]
     pub(crate) fn dictionary_arena_allocation_count(&self) -> usize {
-        self.strings.arena_allocation_count
+        self.strings.arena_allocation_count()
     }
 
     #[cfg(test)]
     pub(crate) fn dictionary_arena_large_allocation_count(&self) -> usize {
-        self.strings.arena_large_allocation_count
+        self.strings.arena_large_allocation_count()
     }
 }
 
@@ -316,220 +285,8 @@ impl MaterializedTrackedStateExactBatch {
     }
 }
 
-const NO_STRING_ORDINAL: u32 = u32::MAX;
-const SMALL_STRING_DICTIONARY_LIMIT: usize = 32;
-const SMALL_STRING_ARENA_BYTES: usize = 1024;
-#[cfg(test)]
-const LARGE_STRING_ALLOCATION_BYTES: usize = 32 * 1024;
-
-enum TrackedStateStringLookup {
-    Small,
-    Hashed(HashMap<u64, u32, RandomState>),
-}
-
-/// Arena-first dictionary builder for historical row identities.
-///
-/// Hash buckets retain compact ordinals into `ranges`; collision chains are
-/// another flat ordinal column. No distinct schema/file value owns a `String`,
-/// and finishing transfers the UTF-8 arena into `Bytes` without recopying it.
-struct TrackedStateStringDictionaryBuilder {
-    bytes: Vec<u8>,
-    ranges: Vec<Range<u32>>,
-    collision_next: Vec<u32>,
-    lookup: TrackedStateStringLookup,
-    hash_builder: RandomState,
-    expected_entry_capacity: usize,
-    maximum_entry_capacity: usize,
-    max_string_len: usize,
-    exact_byte_capacity: bool,
-    #[cfg(test)]
-    arena_allocation_count: usize,
-    #[cfg(test)]
-    arena_large_allocation_count: usize,
-}
-
-impl TrackedStateStringDictionaryBuilder {
-    fn with_capacity(entry_capacity: usize, byte_capacity: usize) -> Self {
-        let expected_entry_capacity = entry_capacity.max(1);
-        Self {
-            bytes: Vec::with_capacity(byte_capacity),
-            ranges: Vec::with_capacity(entry_capacity),
-            collision_next: Vec::with_capacity(entry_capacity),
-            lookup: TrackedStateStringLookup::Small,
-            hash_builder: RandomState::with_seeds(0, 0, 0, 0),
-            expected_entry_capacity,
-            maximum_entry_capacity: expected_entry_capacity,
-            max_string_len: 0,
-            exact_byte_capacity: byte_capacity != 0,
-            #[cfg(test)]
-            arena_allocation_count: usize::from(byte_capacity != 0),
-            #[cfg(test)]
-            arena_large_allocation_count: usize::from(
-                byte_capacity >= LARGE_STRING_ALLOCATION_BYTES,
-            ),
-        }
-    }
-
-    fn intern(&mut self, value: &str) -> u32 {
-        if !matches!(&self.lookup, TrackedStateStringLookup::Small) {
-            return self.intern_hashed(value);
-        }
-        if let Some(ordinal) = self.find_linear(value) {
-            return ordinal;
-        }
-        if self.ranges.len() == SMALL_STRING_DICTIONARY_LIMIT {
-            self.promote_to_hashed(value.len());
-            self.intern_hashed(value)
-        } else {
-            self.append_small(value)
-        }
-    }
-
-    fn find_linear(&self, value: &str) -> Option<u32> {
-        self.ranges
-            .iter()
-            .position(|range| {
-                &self.bytes[range.start as usize..range.end as usize] == value.as_bytes()
-            })
-            .map(|ordinal| {
-                u32::try_from(ordinal).expect("tracked-state string dictionary ordinal exceeds u32")
-            })
-    }
-
-    fn intern_hashed(&mut self, value: &str) -> u32 {
-        let hash = self.hash_builder.hash_one(value.as_bytes());
-        let mut candidate = match &self.lookup {
-            TrackedStateStringLookup::Small => {
-                unreachable!("hashed tracked-state dictionary must be promoted")
-            }
-            TrackedStateStringLookup::Hashed(lookup) => lookup.get(&hash).copied(),
-        };
-        while let Some(ordinal) = candidate {
-            let range = &self.ranges[ordinal as usize];
-            if &self.bytes[range.start as usize..range.end as usize] == value.as_bytes() {
-                return ordinal;
-            }
-            let next = self.collision_next[ordinal as usize];
-            candidate = (next != NO_STRING_ORDINAL).then_some(next);
-        }
-
-        self.append_hashed(value, hash)
-    }
-
-    fn append_small(&mut self, value: &str) -> u32 {
-        let ordinal = self.append_bytes(value);
-        self.collision_next.push(NO_STRING_ORDINAL);
-        ordinal
-    }
-
-    fn append_hashed(&mut self, value: &str, hash: u64) -> u32 {
-        let previous_head = match &self.lookup {
-            TrackedStateStringLookup::Small => {
-                unreachable!("hashed tracked-state dictionary must be promoted")
-            }
-            TrackedStateStringLookup::Hashed(lookup) => {
-                lookup.get(&hash).copied().unwrap_or(NO_STRING_ORDINAL)
-            }
-        };
-        let ordinal = self.append_bytes(value);
-        self.collision_next.push(previous_head);
-        let TrackedStateStringLookup::Hashed(lookup) = &mut self.lookup else {
-            unreachable!("hashed tracked-state dictionary must remain promoted")
-        };
-        lookup.insert(hash, ordinal);
-        ordinal
-    }
-
-    fn append_bytes(&mut self, value: &str) -> u32 {
-        self.max_string_len = self.max_string_len.max(value.len());
-        let required = self
-            .bytes
-            .len()
-            .checked_add(value.len())
-            .expect("tracked-state string dictionary byte count overflow");
-        self.ensure_arena_capacity(required);
-        let start = u32::try_from(self.bytes.len())
-            .expect("tracked-state string dictionary exceeds u32 bytes");
-        self.bytes.extend_from_slice(value.as_bytes());
-        let end = u32::try_from(self.bytes.len())
-            .expect("tracked-state string dictionary exceeds u32 bytes");
-        let ordinal = u32::try_from(self.ranges.len())
-            .expect("tracked-state string dictionary exceeds u32 entries");
-        assert_ne!(
-            ordinal, NO_STRING_ORDINAL,
-            "tracked-state string dictionary reserves the terminal u32 ordinal"
-        );
-        self.ranges.push(start..end);
-        ordinal
-    }
-
-    fn ensure_arena_capacity(&mut self, required: usize) {
-        if required <= self.bytes.capacity() {
-            return;
-        }
-        let projected = match &self.lookup {
-            TrackedStateStringLookup::Small => SMALL_STRING_ARENA_BYTES,
-            TrackedStateStringLookup::Hashed(_) => self
-                .maximum_entry_capacity
-                .saturating_mul(self.max_string_len),
-        };
-        let target = required.max(projected);
-        self.bytes.reserve_exact(target - self.bytes.len());
-        #[cfg(test)]
-        {
-            self.arena_allocation_count += 1;
-            self.arena_large_allocation_count +=
-                usize::from(target >= LARGE_STRING_ALLOCATION_BYTES);
-        }
-    }
-
-    fn promote_to_hashed(&mut self, incoming_len: usize) {
-        self.max_string_len = self.max_string_len.max(incoming_len);
-        let projected_entries = self
-            .expected_entry_capacity
-            .max(self.ranges.len().saturating_add(1));
-        let projected_bytes = projected_entries.saturating_mul(self.max_string_len);
-        if !self.exact_byte_capacity && projected_bytes > self.bytes.capacity() {
-            self.bytes.reserve_exact(projected_bytes - self.bytes.len());
-            #[cfg(test)]
-            {
-                self.arena_allocation_count += 1;
-                self.arena_large_allocation_count +=
-                    usize::from(projected_bytes >= LARGE_STRING_ALLOCATION_BYTES);
-            }
-        }
-
-        let mut lookup = HashMap::with_capacity_and_hasher(
-            projected_entries,
-            RandomState::with_seeds(0, 0, 0, 1),
-        );
-        for ordinal in 0..self.ranges.len() {
-            let ordinal = u32::try_from(ordinal)
-                .expect("tracked-state string dictionary ordinal exceeds u32");
-            let range = &self.ranges[ordinal as usize];
-            let hash = self
-                .hash_builder
-                .hash_one(&self.bytes[range.start as usize..range.end as usize]);
-            self.collision_next[ordinal as usize] =
-                lookup.insert(hash, ordinal).unwrap_or(NO_STRING_ORDINAL);
-        }
-        self.lookup = TrackedStateStringLookup::Hashed(lookup);
-    }
-
-    fn finish(self) -> TrackedStateStringDictionary {
-        TrackedStateStringDictionary {
-            bytes: Bytes::from(self.bytes),
-            ranges: self.ranges,
-            #[cfg(test)]
-            arena_allocation_count: self.arena_allocation_count,
-            #[cfg(test)]
-            arena_large_allocation_count: self.arena_large_allocation_count,
-        }
-    }
-}
-
 struct MaterializedTrackedStateBatchBuilder {
-    strings: TrackedStateStringDictionaryBuilder,
+    strings: StringDictionaryBuilder,
     rows: Vec<MaterializedTrackedStateDescriptor>,
 }
 
@@ -545,9 +302,13 @@ impl MaterializedTrackedStateBatchBuilder {
         dictionary_byte_capacity: usize,
     ) -> Self {
         Self {
-            strings: TrackedStateStringDictionaryBuilder::with_capacity(
+            // The canonical plane sizes its dictionary directly, so it has no
+            // separate rows × identity-columns projection to offer.
+            strings: StringDictionaryBuilder::with_capacity(
+                0,
                 dictionary_entry_capacity,
                 dictionary_byte_capacity,
+                dictionary_byte_capacity != 0,
             ),
             rows: Vec::with_capacity(row_count),
         }
